+++ libnavui-maps/src/main/java/com/mapbox/navigation/ui/maps/camera/data/ViewportDataSourceUpdateObserver.kt
@@ -1,5 +1,7 @@
 package com.mapbox.navigation.ui.maps.camera.data
 
+import androidx.annotation.UiThread
+
 /**
  * Observer that gets notified whenever [ViewportData] changes.
  */
@@ -9,5 +11,6 @@ fun interface ViewportDataSourceUpdateObserver {
      * Called whenever [ViewportData] changes.
      * @param viewportData latest data
      */
+    @UiThread
     fun viewportDataSourceUpdated(viewportData: ViewportData)
 }